Firebase config rile relocation
