+++ lib/contexts/NotificationContext.tsx
@@ -22,7 +22,7 @@ import {
   serverTimestamp,
   Timestamp,
 } from "firebase/firestore";
-import { db } from "@/lib/firebase";
+import { db } from "@/lib/configs/firebase";
 import { useAuth } from "@/lib/contexts/AuthContext";
 import type { Cat } from "@/lib/data/mockData";
